refactor: move -i into entry
When looking at prettier-mirror, I noticed that the entry field had the non-changeable options in it as well, like `--write`. Clang-format's equivalent of write is `-i`, and the hook would not do anything interesting without it, so maybe it should be set that way in this one too? One downside is that clang-format explicitly checks, so this will provide an error if anyone is already overriding args and including `-i`:

```
clang-format: for the -i option: may only occur zero or one times!
```